+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## ??/??/2020 - v0.9.0.0
+
+* (Add) Multi-OS with Linux support
+* (Add) Themes support
+* (Add) Fullscreen support (F11)
+* (Change) GUI was rewritten from Windows Forms to WPF Avalonia, C#
+* (Improvement) More detailed descriptions on error messages
+* (Improvement) GUI is now scalable
+* (Fix) Many bug found and fixed during convertion
+
 ## ??/09/2020 - v0.8.2.5
 
 * (Fix) Extract: Use trail zeros to layer filenames